+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/MoveCancelService.java
@@ -77,7 +77,9 @@ public class MoveCancelService {
 
     try {
 
-      if (accountConfigService.getAccountConfig(move.getCompany()).getAllowRemovalValidatedMove()) {
+      if (paymentMove.getStatusSelect() == MoveRepository.STATUS_DAYBOOK 
+    		  || (paymentMove.getStatusSelect() == MoveRepository.STATUS_VALIDATED 
+    		  && accountConfigService.getAccountConfig(move.getCompany()).getAllowRemovalValidatedMove())) {
         moveRepository.remove(move);
       } else {
         move.setStatusSelect(MoveRepository.STATUS_CANCELED);